move gke functionality to provider gke package
