+++ pkg/handler/v2/external_cluster/upgrade.go
@@ -28,6 +28,7 @@ import (
 	"k8c.io/kubermatic/v2/pkg/handler/v1/common"
 	"k8c.io/kubermatic/v2/pkg/provider"
 	"k8c.io/kubermatic/v2/pkg/provider/cloud/azure"
+	"k8c.io/kubermatic/v2/pkg/provider/cloud/gke"
 	"k8c.io/kubermatic/v2/pkg/resources"
 	"k8c.io/kubermatic/v2/pkg/util/errors"
 )
@@ -66,7 +67,7 @@ func GetUpgradesEndpoint(userInfoGetter provider.UserInfoGetter, projectProvider
 			if err != nil {
 				return nil, err
 			}
-			return providercommon.ListGKEUpgrades(ctx, sa, cloud.GKE.Zone, cloud.GKE.Name)
+			return gke.ListGKEUpgrades(ctx, sa, cloud.GKE.Zone, cloud.GKE.Name)
 		}
 		if cloud.AKS != nil {
 			cred, err := azure.GetCredentialsForAKSCluster(*cloud, secretKeySelector)
@@ -110,7 +111,7 @@ func GetMachineDeploymentUpgradesEndpoint(userInfoGetter provider.UserInfoGetter
 				if err != nil {
 					return nil, err
 				}
-				return providercommon.ListGKEMachineDeploymentUpgrades(ctx, sa, cloud.GKE.Zone, cloud.GKE.Name, req.MachineDeploymentID)
+				return gke.ListGKEMachineDeploymentUpgrades(ctx, sa, cloud.GKE.Zone, cloud.GKE.Name, req.MachineDeploymentID)
 			}
 			if cloud.AKS != nil {
 				cred, err := azure.GetCredentialsForAKSCluster(*cloud, secretKeySelector)
